+++ test/OpenTelemetry.Exporter.OneCollector.Tests/ExtensionFieldInformationManagerTests.cs
@@ -28,6 +28,7 @@ public class ExtensionFieldInformationManagerTests
         var result = extensionFieldInformationManager.TryResolveExtensionFieldInformation("ext.something.fieldName1", out var fieldInformation);
 
         Assert.True(result);
+        Assert.NotNull(fieldInformation);
         Assert.Equal("something", fieldInformation.ExtensionName);
         Assert.Equal("fieldName1", fieldInformation.FieldName);
 
@@ -38,6 +39,7 @@ public class ExtensionFieldInformationManagerTests
         Assert.Equal(1, extensionFieldInformationManager.CountOfCachedExtensionFields);
 
         Assert.True(result);
+        Assert.NotNull(fieldInformation);
         Assert.Equal("something", fieldInformation.ExtensionName);
         Assert.Equal("fieldName1", fieldInformation.FieldName);
 
@@ -46,8 +48,18 @@ public class ExtensionFieldInformationManagerTests
         Assert.Equal(2, extensionFieldInformationManager.CountOfCachedExtensionFields);
 
         Assert.True(result);
+        Assert.NotNull(fieldInformation);
         Assert.Equal("something", fieldInformation.ExtensionName);
         Assert.Equal("field.Name2", fieldInformation.FieldName);
+
+        result = extensionFieldInformationManager.TryResolveExtensionFieldInformation("ext.extra", out fieldInformation);
+
+        Assert.Equal(3, extensionFieldInformationManager.CountOfCachedExtensionFields);
+
+        Assert.True(result);
+        Assert.NotNull(fieldInformation);
+        Assert.Equal("extra", fieldInformation.ExtensionName);
+        Assert.Null(fieldInformation.FieldName);
     }
 
     [Fact]
@@ -59,14 +71,16 @@ public class ExtensionFieldInformationManagerTests
 
         Assert.Equal(1, extensionFieldInformationManager.CountOfCachedExtensionFields);
 
-        Assert.False(extensionFieldInformationManager.TryResolveExtensionFieldInformation("ext.", out _));
+        Assert.False(extensionFieldInformationManager.TryResolveExtensionFieldInformation("EXT.", out _));
 
         Assert.Equal(1, extensionFieldInformationManager.CountOfCachedExtensionFields);
 
-        Assert.False(extensionFieldInformationManager.TryResolveExtensionFieldInformation("ext.something", out _));
+        Assert.False(extensionFieldInformationManager.TryResolveExtensionFieldInformation("ext..", out _));
+        Assert.False(extensionFieldInformationManager.TryResolveExtensionFieldInformation("ext..field", out _));
         Assert.False(extensionFieldInformationManager.TryResolveExtensionFieldInformation("ext.something.", out _));
+        Assert.False(extensionFieldInformationManager.TryResolveExtensionFieldInformation("ext.SOMETHING.", out _));
 
-        Assert.Equal(3, extensionFieldInformationManager.CountOfCachedExtensionFields);
+        Assert.Equal(4, extensionFieldInformationManager.CountOfCachedExtensionFields);
     }
 
     [Fact]
@@ -81,6 +95,7 @@ public class ExtensionFieldInformationManagerTests
             var result = extensionFieldInformationManager.TryResolveExtensionFieldInformation($"ext.something.{fieldName}", out var fieldInformation);
 
             Assert.True(result);
+            Assert.NotNull(fieldInformation);
             Assert.Equal("something", fieldInformation.ExtensionName);
             Assert.Equal(fieldName, fieldInformation.FieldName);
         }